pyproject.toml
[tool.poetry]
authors = ["Institutional Data Initiative at Harvard Law School Library"]
description = ""
name = "institutional-books-1-pipeline"
package-mode = false
readme = "README.md"
version = "0.0.1"

[tool.poetry.dependencies]
accelerate = "^1.4.0"
beautifulsoup4 = "^4.13.3"
boto3 = "^1.37.0"
click = "^8.1.8"
datasets = "^3.5.0"
diskcache = "^5.6.3"
humanize = "^4.12.1"
iso639-lang = "^2.6.0"
langchain = "^0.3.19"
more-itertools = "^10.6.0"
morfessor = "^2.0.6"
numpy = "1.26.4"
ocroscope = {git = "https://github.com/Pleias/OCRoscope.git"}
opencv-python = "^4.11.0.86"
orjson = "^3.10.16"
peewee = "^3.17.9"
pillow = "^11.1.0"
polyglot = "^16.7.4"
protobuf = "^5.29.3"
pyfranc = "^0.2.3"
pyicu = "^2.14"
python = ">=3.11,<3.13"
python-dotenv = "^1.0.1"
python-slugify = "^8.0.4"
requests = "^2.32.3"
scikit-learn = "^1.6.1"
sentencepiece = "^0.2.0"
simhash = "^2.1.2"
tiktoken = "^0.9.0"
torch = "^2.6.0"
torchvision = "^0.21.0"
transformers = "^4.49.0"
pandas = "^2.2.3"
ollama = "^0.4.8"
model2vec = {extras = ["train"], version = "^0.5.0"}
lxml = "^5.4.0"
huggingface-hub = "^0.31.2"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 100
